ml_service/pyproject.toml
[tool.poetry]
name = "ml_service"
version = "0.1.0"
description = ""
authors = ["Lee Sutton <leesutton1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.58.0"
newspaper3k = "^0.2.8"
uvicorn = "^0.11.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
